internal: refactoring
	sys->internal

sys package is not supposed to be consumed
by any other packages, renaming it internal achives
that.

	client.Socket -> client.Conn

On the server side of things p9 uses a io.ReadWriteCloser interface
which has a more minimal interface than net.Conn. This makes the
client implementation orthogonal to the server implementation also
makes it possible to use it quic streams instead of sockets.

	ErrSocket -> ConnError

Err prefix is great for variables, but for custom error types
errors pkg uses FooError format.

	errors.Is(err, ErrSocket) # stutters
	errors.Is(err, ConnError) # doesn't stutter
