+++ fsimpl/staticfs/staticfs.go
@@ -21,8 +21,8 @@ import (
 	"strings"
 
 	"github.com/hugelgupf/p9/fsimpl/templatefs"
+	"github.com/hugelgupf/p9/internal/linux"
 	"github.com/hugelgupf/p9/p9"
-	"github.com/hugelgupf/p9/sys/linux"
 )
 
 // Option is a configurator for New.
