+++ fsimpl/templatefs/readonly.go
@@ -1,8 +1,8 @@
 package templatefs
 
 import (
+	"github.com/hugelgupf/p9/internal/linux"
 	"github.com/hugelgupf/p9/p9"
-	"github.com/hugelgupf/p9/sys/linux"
 )
 
 // NotSymlinkFile denies Readlink with EINVAL.
