+++ fsimpl/templatefs/unimplfs.go
@@ -23,8 +23,8 @@
 package templatefs
 
 import (
+	"github.com/hugelgupf/p9/internal/linux"
 	"github.com/hugelgupf/p9/p9"
-	"github.com/hugelgupf/p9/sys/linux"
 )
 
 // NilCloser returns nil for Close.
