+++ internal/errors_linux.go
@@ -1,11 +1,11 @@
 // +build linux
 
-package sys
+package internal
 
 import (
 	"syscall"
 
-	"github.com/hugelgupf/p9/sys/linux"
+	"github.com/hugelgupf/p9/internal/linux"
 )
 
 func sysErrno(err error) linux.Errno {
