+++ internal/linux/errno.go
