+++ internal/stat_windows.go
@@ -1,4 +1,4 @@
-package sys
+package internal
 
 import (
 	"os"
