+++ p9/client_file.go
@@ -21,7 +21,7 @@ import (
 	"runtime"
 	"sync/atomic"
 
-	"github.com/hugelgupf/p9/sys/linux"
+	"github.com/hugelgupf/p9/internal/linux"
 )
 
 // Attach attaches to a server.
