+++ p9/file.go
@@ -15,7 +15,7 @@
 package p9
 
 import (
-	"github.com/hugelgupf/p9/sys/linux"
+	"github.com/hugelgupf/p9/internal/linux"
 )
 
 // Attacher is provided by the server.
